requirements.txt
asgiref==3.5.2
Django==4.1.1
django-crispy-forms==1.12.0
django-rest==0.8.7
djangorestframework==3.12.4
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
